cuembyBackend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-graphql-jwt==0.3.4
executing==1.0.0
graphene==2.1.9
graphene-django==2.15.0
graphene-django-optimizer==0.6.2
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.30
psycopg2==2.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.3.0
python-gnupg==0.4.9
pytz==2022.2
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
text-unidecode==1.3
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
